close gh-15, gh-7, Wallet balance feature and Create Leaderboard endp…

+++ src/main/java/com/rivertech/betgametask/wallet/repository/WalletRepository.java
@@ -3,6 +3,6 @@ package com.rivertech.betgametask.wallet.repository;
 import com.rivertech.betgametask.wallet.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletRepository extends JpaRepository<Wallet, Integer> {
+public interface WalletRepository extends JpaRepository<Wallet, Long> {
 
 }
